Add feature: error handling

+++ app/src/main/java/io/hexlet/spring/controller/PostController.java
@@ -1,11 +1,11 @@
 package io.hexlet.spring.controller;
 
+import io.hexlet.spring.exception.ResourceNotFoundException;
 import io.hexlet.spring.model.Post;
 import io.hexlet.spring.repository.PostRepository;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,8 +17,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 import static io.hexlet.utils.UpdateEntity.updateEntity;
@@ -26,6 +24,7 @@ import static io.hexlet.utils.UpdateEntity.updateEntity;
 @RestController
 @RequestMapping("/api/posts")
 public class PostController {
+    @Autowired
     private PostRepository postRepository;
 
     public PostController(PostRepository postRepository) {
@@ -46,8 +45,10 @@ public class PostController {
         var post = postRepository.findById(longId);
         if (post.isPresent()) {
             return ResponseEntity.ok().body(post.get());
+        } else {
+            throw new ResourceNotFoundException(String.format("Post with id = %s not found", id));
         }
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+//        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
     }
 
     @PostMapping
@@ -59,7 +60,7 @@ public class PostController {
     @PutMapping("/{id}") // Обновление страницы
     public ResponseEntity<Post> update(@PathVariable String id, @Valid @RequestBody Post data) throws NoSuchFieldException, IllegalAccessException {
         var longId = Long.parseLong(id);
-        var findedPost = postRepository.findById(longId).orElseThrow(EntityNotFoundException::new);
+        var findedPost = postRepository.findById(longId).orElseThrow(() -> new ResourceNotFoundException(String.format("Post with id = %s not found", id)));
         updateEntity(findedPost, data);
         postRepository.save(findedPost);
 
@@ -69,6 +70,9 @@ public class PostController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> destroy(@PathVariable String id) {
         var longId = Long.parseLong(id);
+        if (postRepository.findById(longId).isEmpty()) {
+            throw new ResourceNotFoundException(String.format("Post with id = %s not found", id));
+        }
         postRepository.deleteById(longId);
         return ResponseEntity.status(204).build();
     }
